backend/requirements.txt
Django
djangorestframework
psycopg2
django-cors-headers
pillow
